add light version verify before work
