+++ pom.xml
@@ -12,7 +12,7 @@
 
     <groupId>ru.newsystems</groupId>
     <artifactId>NISPro_bot</artifactId>
-    <version>0.1.1-SNAPSHOT</version>
+    <version>0.2.1-SNAPSHOT</version>
     <name>NISPro_bot</name>
     <description>NISPro_bot</description>
